Live migration for non-serial reads
Patch by Ariel Weisberg; Reviewed by David Capwell for CASSANDRA-19439

+++ src/java/org/apache/cassandra/db/MutableDeletionInfo.java
@@ -23,7 +23,8 @@ import java.util.Iterator;
 import com.google.common.base.Objects;
 
 import org.apache.cassandra.config.DatabaseDescriptor;
-import org.apache.cassandra.db.rows.*;
+import org.apache.cassandra.db.rows.EncodingStats;
+import org.apache.cassandra.db.rows.RangeTombstoneMarker;
 import org.apache.cassandra.utils.ObjectSizes;
 import org.apache.cassandra.utils.memory.ByteBufferCloner;
 
@@ -230,7 +231,7 @@ public class MutableDeletionInfo implements DeletionInfo
         return this;
     }
 
-    public DeletionInfo updateAllTimestampAndLocalDeletionTime(long timestamp, int localDeletionTime)
+    public DeletionInfo updateAllTimestampAndLocalDeletionTime(long timestamp, long localDeletionTime)
     {
         if (partitionDeletion.markedForDeleteAt() != Long.MIN_VALUE)
             partitionDeletion = DeletionTime.build(timestamp, localDeletionTime);
